pyproject.toml
[project]
name = "ai-ledger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.37.0",
    "google-genai>=0.2.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "audio-recorder-streamlit>=0.0.10",
    "python-dateutil>=2.9.0.post0",
    "tenacity>=9.1.2",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.2",
]
